+++ tests/testthat/test-mltools.R
@@ -40,11 +40,7 @@ test_that("look ok", {
   expect_type(list4, "list")
   expect_length(list4, 2L)
   expect_s3_class(look(stratlist, which = 1), "ichimoku")
+  expect_null(look(sample_ohlc_data))
+  expect_silent(look(sample_ohlc_data))
   expect_error(look(stratlist, which = 9), regexp = "one of the strategies")
-  expect_error(look(sample_ohlc_data), regexp = "certain object types")
-  df <- structure(list(a = 1:10), class = "data.frame", row.names = 1:10,
-                  oanda = TRUE, instrument = "INS", price = "B", timestamp = "test")
-  list5 <- look(df)
-  expect_type(list5, "list")
-  expect_length(list5, 3)
 })